Add path based analyser
