+++ dependency-analyser/pom.xml
@@ -8,7 +8,7 @@
         <version>1.0-SNAPSHOT</version>
     </parent>
     <groupId>com.redhat.hacbs</groupId>
-    <artifactId>hacbs-container-analyser</artifactId>
+    <artifactId>hacbs-dependency-analyser</artifactId>
     <version>1.0-SNAPSHOT</version>
     <dependencies>
         <dependency>
@@ -16,8 +16,12 @@
             <artifactId>quarkus-rest-client-reactive</artifactId>
         </dependency>
         <dependency>
+            <groupId>com.redhat.hacbs</groupId>
+            <artifactId>hacbs-resource-model</artifactId>
+        </dependency>
+        <dependency>
             <groupId>io.quarkus</groupId>
-            <artifactId>quarkus-jsonp</artifactId>
+            <artifactId>quarkus-picocli</artifactId>
         </dependency>
         <dependency>
             <groupId>io.quarkus</groupId>
