+++ dependency-analyser/src/main/java/com/redhat/hacbs/container/analyser/Main.java
@@ -1,0 +1,9 @@
+package com.redhat.hacbs.container.analyser;
+
+import io.quarkus.picocli.runtime.annotations.TopCommand;
+import picocli.CommandLine;
+
+@TopCommand()
+@CommandLine.Command(subcommands = AnalyseImage.class)
+public class Main {
+}
