+++ dependency-analyser/src/main/resources/application.properties
@@ -1,0 +1,3 @@
+quarkus.container-image.registry=quay.io
+quarkus.container-image.group=redhat-appstudio
+quarkus.container-image.name=hacbs-jvm-dependency-analyser
