--- operator/src/main/java/com/redhat/hacbs/operator/model/v1alpha1/ArtifactBuildRequestSpec.java
@@ -1,4 +1,0 @@
-package com.redhat.hacbs.operator.model.v1alpha1;
-
-public class ArtifactBuildRequestSpec {
-}
