--- operator/src/main/java/com/redhat/hacbs/operator/model/v1alpha1/ComponentSpec.java
@@ -1,5 +1,0 @@
-package com.redhat.hacbs.operator.model.v1alpha1;
-
-public class ComponentSpec {
-
-}
