+++ resource-model/src/main/java/com/redhat/hacbs/resources/model/v1alpha1/ArtifactBuildRequestSpec.java
@@ -1,0 +1,4 @@
+package com.redhat.hacbs.resources.model.v1alpha1;
+
+public class ArtifactBuildRequestSpec {
+}
