+++ resource-model/src/main/java/com/redhat/hacbs/resources/model/v1alpha1/ComponentSpec.java
@@ -1,0 +1,5 @@
+package com.redhat.hacbs.resources.model.v1alpha1;
+
+public class ComponentSpec {
+
+}
