+++ resource-model/src/main/java/com/redhat/hacbs/resources/model/v1alpha1/ModelConstants.java
@@ -1,4 +1,4 @@
-package com.redhat.hacbs.operator.model.v1alpha1;
+package com.redhat.hacbs.resources.model.v1alpha1;
 
 public class ModelConstants {
     public static final String GROUP = "hacbs.redhat.com";
